+++ src/parser/ELR/builder/model/index.ts
@@ -1,5 +1,4 @@
 export * from "./conflict";
-export * from "./context";
 export * from "./definition";
 export * from "./temp-grammar";
 export * from "./resolver";