+++ src/parser/ELR/model/re-lex.ts
@@ -1,6 +1,6 @@
 import { ILexer } from "../../../lexer";
 import { ASTNode } from "../../ast";
-import { Callback, ParserContext } from "./context";
+import { Callback, GrammarRuleContext } from "./context";
 
 export type ReLexStack<State, T> = {
   readonly stateStack: State[];
@@ -13,5 +13,5 @@ export type ReLexStack<State, T> = {
 
 export type RollbackStack<T> = {
   readonly rollback: Callback<T>;
-  readonly context: ParserContext<T>;
+  readonly context: GrammarRuleContext<T>;
 }[];